requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
google-genai>=0.8.0
pdfplumber>=0.11.0
python-docx>=1.1.2
python-multipart>=0.0.9
python-dotenv>=1.0.1
pytest>=8.2.0
httpx>=0.27.0
